+++ test/components/action_list_item_test.exs
@@ -230,9 +230,9 @@ defmodule PrimerLive.TestComponents.ActionListItemTest do
            |> format_html() ==
              """
              <li class="ActionList-item" aria-selected="true" role="option"><span class="ActionList-content"><span class="ActionList-item-visual ActionList-item-visual--leading"><span class="FormControl-checkbox-wrap ActionList-item-multiSelectIcon">
-             <input checked class="FormControl-checkbox" type="checkbox" value="" /></span></span><span class="ActionList-item-label">Item</span></span></li>
+             <input checked class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span class="ActionList-item-label">Item</span></span></li>
              <li class="ActionList-item" aria-selected="false" role="option"><span class="ActionList-content"><span class="ActionList-item-visual ActionList-item-visual--leading"><span class="FormControl-checkbox-wrap ActionList-item-multiSelectIcon">
-             <input class="FormControl-checkbox" type="checkbox" value="" /></span></span><span class="ActionList-item-label">Item</span></span></li>
+             <input class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span class="ActionList-item-label">Item</span></span></li>
              """
              |> format_html()
   end